+++ src/JordanMod/modules/open_mesa/OpenMesaModule.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace JordanMod.Modules.OpenMesa;
+
+[Module(Enabled = true)]
+class OpenMesaModule : Module
+{
+	public override string ModuleName => "Open Mesa Module";
+    
+    public override Type[] GetPatches()
+    {
+        return [typeof(OpenMesaPatch)];
+    }
+}